uncommitted changed from friday

+++ MasterProject/MeasurementVector.cs
@@ -7,14 +7,14 @@ namespace MasterProject
 {
     public class MeasurementVector
     {
-        public int dimension1;
-        public int dimension2;
+        public float dimension1;
+        public float dimension2;
 
         public MeasurementVector()
         {
         }
 
-        public MeasurementVector(int d1, int d2)
+        public MeasurementVector(float d1, float d2)
         {
             dimension1 = d1;
             dimension2 = d2;
